% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{aggregate_to_epiweeks}
\alias{aggregate_to_epiweeks}
\title{aggregate_to_epiweeks}
\usage{
aggregate_to_epiweeks(x)
}
\description{
Aggregate data to epiweek and back-calculate release and inflow using mass balance
}
