% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_parameters_to_release_harmonic}
\alias{convert_parameters_to_release_harmonic}
\title{convert_parameters_to_release_harmonic}
\usage{
convert_parameters_to_release_harmonic(parameters)
}
\arguments{
\item{parameters}{vector of length 4 giving, in order, first sine term, first cosine term, second sine term, second cosine term.}
}
\value{
a tibble of storage target levels by week
}
\description{
fit parameters of a constrained harmonic
}
