% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_parameters_to_storage_targets}
\alias{convert_parameters_to_storage_targets}
\title{convert_parameters_to_storage_targets}
\usage{
convert_parameters_to_storage_targets(
  parameters,
  target_name,
  constrain = TRUE
)
}
\arguments{
\item{parameters}{vector of length 5 giving, in order, intercept, sine term, cosine term, and upper and lower constraints of the harmonic.}

\item{target_name}{optional. Character string naming the target. E.g., "flood" or "conservation." Default is simply "target."}

\item{constrain}{logical. Constrain targets?}
}
\value{
a tibble of storage target levels by week
}
\description{
fit parameters of a constrained harmonic
}
