% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_targets_to_GRanD.R
\name{extrapolate_targets_to_GRanD}
\alias{extrapolate_targets_to_GRanD}
\title{extrapolate_targets_to_GRanD}
\usage{
extrapolate_targets_to_GRanD(USRDATS_path, targets_path)
}
\arguments{
\item{USRDATS_path}{path to USRDATS data}

\item{targets_path}{path to fitted targets (generated using fit_targets())}
}
\value{
tibble dam ids to copy
}
\description{
use an set of inferred storage targets to extrapolate storage targets to all dams in GRanD
}
