% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_closest_dam}
\alias{find_closest_dam}
\title{find_closest_dam}
\usage{
find_closest_dam(dam_attr, other_dams)
}
\arguments{
\item{dam_attr}{attributes of target dam}

\item{other_dams}{table of attributes for possible canditate dams to replicate}
}
\value{
GRAND_ID of the target dam
}
\description{
finds the dam that is closest in terms of purposes served and Euclidean distance
}
