% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit_constrained_harmonic}
\alias{fit_constrained_harmonic}
\title{fit_constrained_harmonic}
\usage{
fit_constrained_harmonic(data_for_harmonic_fitting)
}
\arguments{
\item{data_for_harmonic_fitting}{tibble with headers epiweek and s_pct}
}
\value{
optimization output; a list
}
\description{
fit parameters of a constrained harmonic
}
