% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_release_function.R
\name{fit_release_function}
\alias{fit_release_function}
\title{fit_release_function}
\usage{
fit_release_function(USRDATS_path, GRanD_path, dam_id, targets_path)
}
\arguments{
\item{USRDATS_path}{path to USRDATS data}

\item{GRanD_path}{path to v1.3 of GRanD database}

\item{dam_id}{integer id of dam; same as GRanD ID}

\item{targets_path}{path to fitted targets. If NULL, fit_targets() will be run.}
}
\value{
tibble of observed dam data (storage, inflow, release)
}
\description{
fit parameters of weekly-varying release function
}
