% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_targets.R
\name{fit_targets}
\alias{fit_targets}
\title{fit_targets}
\usage{
fit_targets(USRDATS_path, dam_id)
}
\arguments{
\item{USRDATS_path}{path to USRDATS data}

\item{dam_id}{integer id of dam; same as GRanD ID}
}
\value{
tibble of observed dam data (storage, inflow, release)
}
\description{
fit parameters of storage targets
}
