% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{read_reservoir_attributes}
\alias{read_reservoir_attributes}
\title{read_reservoir_attributes}
\usage{
read_reservoir_attributes(USRDATS_path, dam_id = NULL)
}
\arguments{
\item{USRDATS_path}{directory containing reservoir input time series}

\item{dam_id}{integer id of dam; same as GRanD ID. If NULL, all attributes are returned.}
}
\value{
tibble of reservoir attributes for selected dams
}
\description{
reads reservoir time series data
}
