% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{read_reservoir_data}
\alias{read_reservoir_data}
\title{read_reservoir_data}
\usage{
read_reservoir_data(USRDATS_path, dam_id)
}
\arguments{
\item{USRDATS_path}{directory containing reservoir input time series}

\item{dam_id}{integer id of dam; same as GRanD ID}
}
\value{
tibble of observed dam data (storage, inflow, release)
}
\description{
reads raw reservoir time series data
}
