% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starfit-package.R
\docType{package}
\name{starfit-package}
\alias{starfit}
\alias{starfit-package}
\title{starfit: Storage Target and Release Function Inference Tool}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A tool for inferring reservoir storage targets and release functions.
}
\author{
\strong{Maintainer}: Sean Turner \email{sean.turner@pnnl.gov} (\href{https://orcid.org/0000-0003-4400-9800}{ORCID})

}
\keyword{internal}
