% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_targets_to_GRanD.R
\name{extrapolate_targets_to_GRanD}
\alias{extrapolate_targets_to_GRanD}
\title{extrapolate_targets_to_GRanD}
\usage{
extrapolate_targets_to_GRanD(
  GRanD_path,
  targets_path,
  include_all = FALSE,
  HUC_04_correction = TRUE,
  distance_only = FALSE
)
}
\arguments{
\item{GRanD_path}{path to v1.3 of GRanD database}

\item{targets_path}{path to fitted targets (generated using fit_targets())}

\item{include_all}{T/F (if T, returns results for dams in the trained set)}

\item{HUC_04_correction}{T/F (if T, replaces Great Lakes HUC2 with neighboring HUC2s...(deals with missing data for all of Great Lakes))}
}
\value{
tibble dam ids to copy
}
\description{
use an set of inferred storage targets to extrapolate storage targets to all dams in GRanD
}
