% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulecurve-package.R
\docType{package}
\name{rulecurve-package}
\alias{rulecurve}
\alias{rulecurve-package}
\title{rulecurve: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A tool for inferring reservoir storage targets and release rules.
}
\author{
\strong{Maintainer}: Sean Turner \email{sean.turner@pnnl.gov} (\href{https://orcid.org/0000-0003-4400-9800}{ORCID})

}
\keyword{internal}
